Update lib_main.c
Made a search function and description function. The checkout code is implemented but does not work. It is a part of the search function and implemented identically in a couple of different places. It is implement after asking for a book's description and implemented as a separate option, both for search results where only one book shows up as well as for search results where two or more results pop up. The return function is not implemented at all.